requirements.txt
Pillow==7.0.0
opencv-python==4.7.0.72
glob2==0.7
numpy>=1.18.1